+++ src/lib/_helpers/diff.ts
@@ -131,10 +131,8 @@ export interface DiffAction {
 type IndexValue = [number, number];
 
 function nearestEnd<T>(src: Slice<T>, target: Slice<T>): IndexValue {
-  const xSeq: Sequence<T> = src.sequence;
   const x1: number = 0;
   const x2: number = src.length;
-  const ySeq: Sequence<T> = target.sequence;
   const y1: number = 0;
   const y2: number = target.length;
   const xLen: number = x2 - x1;