+++ src/lib/types/tagged-union.ts
@@ -3,7 +3,6 @@ import { lazyProperties } from "../_helpers/lazy-properties";
 import { IoType, Lazy, Reader, VersionedType, Writer } from "../core";
 import { createInvalidTypeError } from "../errors/invalid-type";
 import { createLazyOptionsError } from "../errors/lazy-options";
-import { createNotImplementedError } from "../errors/not-implemented";
 import { readVisitor } from "../readers/read-visitor";
 import { testError } from "../test-error";
 import { DocumentType } from "./document";
@@ -14,8 +13,8 @@ export type Name = "union";
 export const name: Name = "union";
 export type Diff = any;
 
-export interface TaggedUnionTypeOptions<T extends {}, M extends DocumentType<T> = DocumentType<T>> {
-  variants: M[];
+export interface TaggedUnionTypeOptions<T, K extends DocumentType<T> = DocumentType<T>> {
+  variants: ReadonlyArray<K>;
   tag: keyof T;
 }
 
@@ -23,21 +22,21 @@ export type TestWithVariantResult<T> =
   [true, VersionedType<T, any>]
   | [false, VersionedType<T, any> | undefined];
 
-export class TaggedUnionType<T extends {}, M extends DocumentType<T> = DocumentType<T>> implements IoType<T>,
-  TaggedUnionTypeOptions<T, M> {
+export class TaggedUnionType<T, K extends DocumentType<T> = DocumentType<T>> implements IoType<T>,
+  TaggedUnionTypeOptions<T, K> {
   readonly name: Name = name;
-  readonly variants: M[];
+  readonly variants: ReadonlyArray<K>;
   readonly tag: keyof T;
 
-  private _options?: Lazy<TaggedUnionTypeOptions<T, M>>;
+  private _options?: Lazy<TaggedUnionTypeOptions<T, K>>;
 
   private _outTag: string | undefined;
 
   private _tagType: TsEnumType<any> | undefined;
 
-  private _valueToVariantMap: Map<any, M> | undefined;
+  private _valueToVariantMap: Map<any, K> | undefined;
 
-  constructor(options: Lazy<TaggedUnionTypeOptions<T, M>>) {
+  constructor(options: Lazy<TaggedUnionTypeOptions<T, K>>) {
     // TODO: Remove once TS 2.7 is better supported by editors
     this.variants = <any> undefined;
     this.tag = <any> undefined;
@@ -54,14 +53,14 @@ export class TaggedUnionType<T extends {}, M extends DocumentType<T> = DocumentT
     }
   }
 
-  match(value: T): M | undefined {
+  match(value: T): K | undefined {
     const tag: keyof T = this.tag;
     const tagValue: any = value[tag];
     if (tagValue === undefined) {
       return undefined;
       // throw new Incident("MissingTag", {union: this, value});
     }
-    const variant: M | undefined = this.getValueToVariantMap().get(tagValue); // tagToVariant
+    const variant: K | undefined = this.getValueToVariantMap().get(tagValue); // tagToVariant
     if (variant === undefined) {
       return undefined;
       // throw new Incident("VariantNotFound", {union: this, value});
@@ -69,12 +68,12 @@ export class TaggedUnionType<T extends {}, M extends DocumentType<T> = DocumentT
     return variant;
   }
 
-  matchTrusted(value: T): M {
+  matchTrusted(value: T): K {
     return this.match(value)!;
   }
 
   write<W>(writer: Writer<W>, value: T): W {
-    const variant: M | undefined = this.match(value);
+    const variant: K | undefined = this.match(value);
     if (variant === undefined) {
       throw new Incident("VariantNotFound", {union: this, value});
     }
@@ -88,20 +87,26 @@ export class TaggedUnionType<T extends {}, M extends DocumentType<T> = DocumentT
     return this.variantRead(reader, raw)[1];
   }
 
-  variantRead<R>(reader: Reader<R>, raw: R): [M, T] {
+  variantRead<R>(reader: Reader<R>, raw: R): [K, T] {
     return reader.readDocument(raw, readVisitor({
-      fromMap: <RK, RV>(input: Map<RK, RV>, keyReader: Reader<RK>, valueReader: Reader<RV>): [M, T] => {
+      fromMap: <RK, RV>(input: Map<RK, RV>, keyReader: Reader<RK>, valueReader: Reader<RV>): [K, T] => {
         const outTag: string = this.getOutTag();
-        if (!input.has(outTag as any)) { // TODO: remove cast
-          throw new Incident("MissingOutTag");
+        for (const [rawKey, rawValue] of input) {
+          const outKey: string = keyReader.readString(
+            rawKey,
+            readVisitor({fromString: (input: string): string  => input}),
+          );
+          if (outKey !== outTag) {
+            continue;
+          }
+          const tagValue: any = this.getTagType().read(valueReader, rawValue);
+          const variant: K | undefined = this.getValueToVariantMap().get(tagValue); // tagToVariant
+          if (variant === undefined) {
+            throw new Incident("VariantNotFound", {union: this, tagValue});
+          }
+          return [variant, variant.read!(reader, raw)];
         }
-        const outTagRawValue: any = input.get(outTag as any); // TODO: remove cast
-        const tagValue: any = this.getTagType().read(valueReader, outTagRawValue);
-        const variant: M | undefined = this.getValueToVariantMap().get(tagValue); // tagToVariant
-        if (variant === undefined) {
-          throw new Incident("VariantNotFound", {union: this, tagValue});
-        }
-        return [variant, variant.read!(reader, raw)];
+        throw new Incident("MissingOutTag");
       },
     }));
   }
@@ -110,7 +115,7 @@ export class TaggedUnionType<T extends {}, M extends DocumentType<T> = DocumentT
     if (typeof value !== "object" || value === null) {
       return createInvalidTypeError("object", value);
     }
-    const variant: M | undefined = this.match(value);
+    const variant: K | undefined = this.match(value);
     if (variant === undefined) {
       return new Incident("UnknownUnionVariant", "Unknown union variant");
     }
@@ -118,7 +123,7 @@ export class TaggedUnionType<T extends {}, M extends DocumentType<T> = DocumentT
   }
 
   // testWithVariant(val: T): TestWithVariantResult<T> {
-  //   const variant: M | undefined = this.match(val);
+  //   const variant: K | undefined = this.match(val);
   //   if (variant === undefined) {
   //     return [false as false, undefined];
   //   }
@@ -129,7 +134,7 @@ export class TaggedUnionType<T extends {}, M extends DocumentType<T> = DocumentT
     if (typeof value !== "object" || value === null) {
       return false;
     }
-    const type: M | undefined = this.match(value);
+    const type: K | undefined = this.match(value);
     if (type === undefined) {
       return false;
     }
@@ -138,8 +143,8 @@ export class TaggedUnionType<T extends {}, M extends DocumentType<T> = DocumentT
 
   // TODO: Always return true?
   equals(val1: T, val2: T): boolean {
-    const type1: M = this.matchTrusted(val1);
-    const type2: M = this.matchTrusted(val2);
+    const type1: K = this.matchTrusted(val1);
+    const type2: K = this.matchTrusted(val2);
     return type1 === type2 && type1.equals(val1, val2);
   }
 
@@ -151,11 +156,11 @@ export class TaggedUnionType<T extends {}, M extends DocumentType<T> = DocumentT
     if (this._options === undefined) {
       throw createLazyOptionsError(this);
     }
-    const options: TaggedUnionTypeOptions<T, M> = typeof this._options === "function"
+    const options: TaggedUnionTypeOptions<T, K> = typeof this._options === "function"
       ? this._options()
       : this._options;
     delete this._options;
-    const variants: M[] = options.variants;
+    const variants: ReadonlyArray<K> = options.variants;
     const tag: keyof T = options.tag;
     Object.assign(this, {variants, tag});
   }
@@ -168,7 +173,7 @@ export class TaggedUnionType<T extends {}, M extends DocumentType<T> = DocumentT
       const tag: keyof T = this.tag;
       let outTag: string | undefined = undefined;
       for (const variant of this.variants) {
-        const cur: string = variant.getOutKey(tag);
+        const cur: string = variant.getOutKey(tag as any);
         if (outTag === undefined) {
           outTag = cur;
         } else if (cur !== outTag) {
@@ -204,10 +209,10 @@ export class TaggedUnionType<T extends {}, M extends DocumentType<T> = DocumentT
     return this._tagType;
   }
 
-  private getValueToVariantMap(): Map<any, M> {
+  private getValueToVariantMap(): Map<any, K> {
     if (this._valueToVariantMap === undefined) {
       const tag: keyof T = this.tag;
-      const valueToVariantMap: Map<any, M> = new Map();
+      const valueToVariantMap: Map<any, K> = new Map();
       for (const variant of this.variants) {
         const lit: LiteralType<any> = variant.properties[tag].type as any;
         if (valueToVariantMap.has(lit.value)) {