+++ src/test/types/any.spec.ts
@@ -1,4 +1,4 @@
-import { assert } from "chai";
+import chai from "chai";
 import { JsonReader } from "../../lib/readers/json";
 import { JsonValueReader } from "../../lib/readers/json-value";
 import { AnyType } from "../../lib/types/any";
@@ -9,8 +9,8 @@ describe("AnyType", function () {
     it("should read the expected top-level values", function () {
       const reader: JsonReader = new JsonReader();
       const $Any: AnyType = new AnyType();
-      assert.deepEqual($Any.read(reader, "0"), "0");
-      assert.deepEqual($Any.read(reader, "{\"foo\": \"bar\""), "{\"foo\": \"bar\"");
+      chai.assert.deepEqual($Any.read(reader, "0"), "0");
+      chai.assert.deepEqual($Any.read(reader, "{\"foo\": \"bar\""), "{\"foo\": \"bar\"");
     });
     it("should read the expected nested values", function () {
       const reader: JsonReader = new JsonReader();
@@ -24,7 +24,7 @@ describe("AnyType", function () {
         properties: {foo: {type: $Any}},
       });
 
-      assert.deepEqual($FooBarQuz.read(reader, "{\"foo\": {\"bar\": \"quz\"}}"), {foo: {bar: "quz"}});
+      chai.assert.deepEqual($FooBarQuz.read(reader, "{\"foo\": {\"bar\": \"quz\"}}"), {foo: {bar: "quz"}});
     });
   });
 
@@ -32,8 +32,8 @@ describe("AnyType", function () {
     it("should read the expected values", function () {
       const reader: JsonValueReader = new JsonValueReader();
       const $Any: AnyType = new AnyType();
-      assert.deepEqual($Any.read(reader, 0), 0);
-      assert.deepEqual($Any.read(reader, {foo: "bar"}), {foo: "bar"});
+      chai.assert.deepEqual($Any.read(reader, 0), 0);
+      chai.assert.deepEqual($Any.read(reader, {foo: "bar"}), {foo: "bar"});
     });
   });
 });