+++ src/test/types/codepoint-string.spec.ts
@@ -1,10 +1,11 @@
-import { assert } from "chai";
+import chai from "chai";
+import unorm from "unorm";
 import { CodepointStringType } from "../../lib/types/codepoint-string";
 import { runTests, TypedValue } from "../helpers/test";
 
 describe("CodepointStringType", function () {
   describe("basic support", function () {
-    const type: CodepointStringType = new CodepointStringType({maxCodepoints: 500});
+    const type: CodepointStringType = new CodepointStringType({maxCodepoints: 500, unorm});
 
     const items: TypedValue[] = [
       // Valid items
@@ -34,32 +35,32 @@ describe("CodepointStringType", function () {
 
   describe("Ensure valid codepoints with Javascript (UCS2) strings", function () {
     it("should accept the empty string, when requiring length exactly 0", function () {
-      assert.isTrue(new CodepointStringType({minCodepoints: 0, maxCodepoints: 0}).test(""));
+      chai.assert.isTrue(new CodepointStringType({minCodepoints: 0, maxCodepoints: 0, unorm}).test(""));
     });
     it("should accept the string \"a\" (ASCII codepoint), when requiring length exactly 1", function () {
-      assert.isTrue(new CodepointStringType({minCodepoints: 1, maxCodepoints: 1}).test("a"));
+      chai.assert.isTrue(new CodepointStringType({minCodepoints: 1, maxCodepoints: 1, unorm}).test("a"));
     });
     it("should accept the string \"∑\" (BMP codepoint), when requiring length exactly 1", function () {
-      assert.isTrue(new CodepointStringType({minCodepoints: 1, maxCodepoints: 1}).test("∑"));
+      chai.assert.isTrue(new CodepointStringType({minCodepoints: 1, maxCodepoints: 1, unorm}).test("∑"));
     });
     it("should reject the string \"𝄞\" (non-BMP codepoint), when requiring length exactly 2", function () {
-      assert.isFalse(new CodepointStringType({minCodepoints: 2, maxCodepoints: 2}).test("𝄞"));
+      chai.assert.isFalse(new CodepointStringType({minCodepoints: 2, maxCodepoints: 2, unorm}).test("𝄞"));
     });
     it("should accept the string \"𝄞\" (non-BMP codepoint), when requiring length exactly 1", function () {
-      assert.isTrue(new CodepointStringType({minCodepoints: 1, maxCodepoints: 1}).test("𝄞"));
+      chai.assert.isTrue(new CodepointStringType({minCodepoints: 1, maxCodepoints: 1, unorm}).test("𝄞"));
     });
     describe("should reject unmatched surrogate halves", function () {
       // 𝄞 corresponds to the surrogate pair (0xd834, 0xdd1e)
-      const type: CodepointStringType = new CodepointStringType({maxCodepoints: 500});
+      const type: CodepointStringType = new CodepointStringType({maxCodepoints: 500, unorm});
       const items: string[] = ["\ud834", "a\ud834", "\ud834b", "a\ud834b", "\udd1e", "a\udd1e", "\udd1eb", "a\udd1eb"];
       for (const item of items) {
         it(JSON.stringify(item), function () {
-          assert.isFalse(type.test(item));
+          chai.assert.isFalse(type.test(item));
         });
       }
     });
     it("should reject reversed (invalid) surrogate pairs", function () {
-      assert.isFalse(new CodepointStringType({maxCodepoints: 500}).test("\udd1e\ud834"));
+      chai.assert.isFalse(new CodepointStringType({maxCodepoints: 500, unorm}).test("\udd1e\ud834"));
     });
   });
 });