+++ src/test/types/ucs2-string.spec.ts
@@ -1,4 +1,4 @@
-import { assert } from "chai";
+import chai from "chai";
 import { Ucs2StringType } from "../../lib/types/ucs2-string";
 import { runTests, TypedValue } from "../helpers/test";
 
@@ -34,19 +34,19 @@ describe("Ucs2StringType", function () {
 
   describe("Simple UCS2 behavior", function () {
     it("should accept the empty string, when requiring length exactly 0", function () {
-      assert.isTrue(new Ucs2StringType({minLength: 0, maxLength: 0}).test(""));
+      chai.assert.isTrue(new Ucs2StringType({minLength: 0, maxLength: 0}).test(""));
     });
     it("should accept the string \"a\" (ASCII codepoint), when requiring length exactly 1", function () {
-      assert.isTrue(new Ucs2StringType({minLength: 1, maxLength: 1}).test("a"));
+      chai.assert.isTrue(new Ucs2StringType({minLength: 1, maxLength: 1}).test("a"));
     });
     it("should accept the string \"∑\" (BMP codepoint), when requiring length exactly 1", function () {
-      assert.isTrue(new Ucs2StringType({minLength: 1, maxLength: 1}).test("∑"));
+      chai.assert.isTrue(new Ucs2StringType({minLength: 1, maxLength: 1}).test("∑"));
     });
     it("should accept the string \"𝄞\" (non-BMP codepoint), when requiring length exactly 2", function () {
-      assert.isTrue(new Ucs2StringType({minLength: 2, maxLength: 2}).test("𝄞"));
+      chai.assert.isTrue(new Ucs2StringType({minLength: 2, maxLength: 2}).test("𝄞"));
     });
     it("should reject the string \"𝄞\" (non-BMP codepoint), when requiring length exactly 1", function () {
-      assert.isFalse(new Ucs2StringType({minLength: 1, maxLength: 1}).test("𝄞"));
+      chai.assert.isFalse(new Ucs2StringType({minLength: 1, maxLength: 1}).test("𝄞"));
     });
     it("should accept unmatched surrogate halves", function () {
       // 𝄞 corresponds to the surrogate pair (0xd834, 0xdd1e)
@@ -54,12 +54,12 @@ describe("Ucs2StringType", function () {
       const items: string[] = ["\ud834", "a\ud834", "\ud834b", "a\ud834b", "\udd1e", "a\udd1e", "\udd1eb", "a\udd1eb"];
       for (const item of items) {
         it(JSON.stringify(item), function () {
-          assert.isTrue(type.test(item));
+          chai.assert.isTrue(type.test(item));
         });
       }
     });
     it("should accept reversed (invalid) surrogate pairs", function () {
-      assert.isTrue(new Ucs2StringType({maxLength: 500}).test("\udd1e\ud834"));
+      chai.assert.isTrue(new Ucs2StringType({maxLength: 500}).test("\udd1e\ud834"));
     });
   });
 });